Rubocop Changes

+++ app/controllers/members_controller.rb
@@ -6,7 +6,6 @@ class MembersController < ApplicationController
   before_action :member_admin_deletion_protection, only: %i[edit update destroy index]
   before_action :authenticate_user
 
-
   @@sorting = ''
 
   def index
@@ -47,15 +46,17 @@ class MembersController < ApplicationController
                          marketingPoints: params[:member][:marketingPoints],
                          totalPoints: params[:member][:totalPoints],
                          admin_id: params[:member][:admin_id]
-    )
+                        )
     respond_to do |format|
       if @member.save
         # if the member does not have connected account connect email to member
         if @member.admin.nil? && @user.nil?
           if params[:member][:admin_password] == 'Officer'
-            @member.update(admin_id: current_admin.id, position: 'Admin', civicPoints: 0, outreachPoints: 0, socialPoints: 0, marketingPoints: 0, totalPoints: 0, status: "true")
+            @member.update(admin_id: current_admin.id, position: 'Admin', civicPoints: 0, outreachPoints: 0, socialPoints: 0, marketingPoints: 0, 
+                           totalPoints: 0, status: 'true')
           else
-            @member.update(admin_id: current_admin.id, position: 'Member', civicPoints: 0, outreachPoints: 0, socialPoints: 0, marketingPoints: 0, totalPoints: 0)
+            @member.update(admin_id: current_admin.id, position: 'Member', civicPoints: 0, outreachPoints: 0, socialPoints: 0, marketingPoints: 0, 
+                           totalPoints: 0)
           end
         end
         format.html { redirect_to member_url(@member), notice: 'Member was successfully created.' }
@@ -88,9 +89,7 @@ class MembersController < ApplicationController
     @member = Member.find(params[:id])
     @member.destroy
     Admin.all.each do |admin|
-      if admin.member.nil?
-        admin.destroy
-      end
+      admin.destroy if admin.member.nil?
     end
     redirect_to members_path
   end
@@ -108,13 +107,11 @@ class MembersController < ApplicationController
 
       Member.where(status: nil).destroy_all
       Admin.all.each do |admin|
-        if admin.member.nil?
-          admin.destroy
-        end
+        admin.destroy if admin.member.nil?
       end
 
       # Member.where(status: false).destory_all
-      redirect_to members_path, notice: "Members Accepted"
+      redirect_to members_path, notice: 'Members Accepted'
     end
   end
 
@@ -140,9 +137,7 @@ class MembersController < ApplicationController
   end
 
   def member_admin_deletion_protection
-    if @user.nil?
-      redirect_to new_member_path
-    end
+    redirect_to new_member_path if @user.nil?
   end
 
   # only lets admins on certain pages
@@ -152,20 +147,12 @@ class MembersController < ApplicationController
 
   # allows admins to check off on who has access to site
   def authenticate_user
-    unless @user.nil?
-      if @user.status.nil?
-        redirect_to root_path notice: 'Pending Leadership approval'
-      end
-    end
+    redirect_to root_path notice: 'Pending Leadership approval' if !@user.nil? && @user.status.nil?
   end
 
   # allows admins to check off on who has access to site
   def authenticate_user
-    if @user != nil
-      if @user.status == nil
-        redirect_to root_path notice: "Pending Leadership approval"
-      end
-    end
+    redirect_to root_path notice: 'Pending Leadership approval' if !@user.nil? && @user.status.nil?
   end
 
 end
